UPSTREAM: cpuidle: Add 'above' and 'below' idle state metrics
Add two new metrics for CPU idle states, "above" and "below", to count
the number of times the given state had been asked for (or entered
from the kernel's perspective), but the observed idle duration turned
out to be too short or too long for it (respectively).

These metrics help to estimate the quality of the CPU idle governor
in use.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
[ Partially cherry-pick from 04dab58a39d402162a7effe7278df8cd41557252 ]
Signed-off-by: Koushik Dey <kousickdey3@gmail.com>
Change-Id: I2bf8a71994a16be4c0b1314dfbb1bcfe2fc2e432