UPSTREAM: cpuidle: teo: Avoid code duplication in conditionals
There are three places in teo_select() where a given amount of time
is compared with TICK_NSEC if tick_nohz_tick_stopped() returns true,
which is a bit of duplicated code.

Avoid that code duplication by defining a helper function to do the
check and using it in all of the places in question.

No intentional functional impact.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

Change-Id: I23d44b6f77f13a8ed8df3367b5cf3cf9c6379977
Signed-off-by: Koushik Dey <kousickdey3@gmail.com>